// src/types/models/Destination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Destination = {
    Field: 'field',
    CardPool: 'cardPool',
    CardQueue: 'cardQueue',
    Skip: 'skip'
} as const;
export type Destination = typeof Destination[keyof typeof Destination];


export function instanceOfDestination(value: any): boolean {
    for (const key in Destination) {
        if (Object.prototype.hasOwnProperty.call(Destination, key)) {
            if (Destination[key as keyof typeof Destination] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DestinationFromJSON(json: any): Destination {
    return DestinationFromJSONTyped(json, false);
}

export function DestinationFromJSONTyped(json: any, ignoreDiscriminator: boolean): Destination {
    return json as Destination;
}

export function DestinationToJSON(value?: Destination | null): any {
    return value as any;
}

export function DestinationToJSONTyped(value: any, ignoreDiscriminator: boolean): Destination {
    return value as Destination;
}