// src/types/models/PlayReason.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PlayReason = {
    ByChara: 'ByChara',
    ByValue: 'ByValue',
    ByWordPlay: 'ByWordPlay',
    ByWordPlay2: 'ByWordPlay2'
} as const;
export type PlayReason = typeof PlayReason[keyof typeof PlayReason];


export function instanceOfPlayReason(value: any): boolean {
    for (const key in PlayReason) {
        if (Object.prototype.hasOwnProperty.call(PlayReason, key)) {
            if (PlayReason[key as keyof typeof PlayReason] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlayReasonFromJSON(json: any): PlayReason {
    return PlayReasonFromJSONTyped(json, false);
}

export function PlayReasonFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlayReason {
    return json as PlayReason;
}

export function PlayReasonToJSON(value?: PlayReason | null): any {
    return value as any;
}

export function PlayReasonToJSONTyped(value: any, ignoreDiscriminator: boolean): PlayReason {
    return value as PlayReason;
}